pyproject.toml
[tool.poetry]
name = "fasthtml-embeddings-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.3.3"
psycopg2 = "^2.9.9"
torch = "^2.4.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
